Spelina/navigate to draft editing if it already exists (#2874)
* navigating to draft editing if it already exists

* removed comment, corrected actions order

* small improvement of subscription

* accessibility improvements for workshop cards

* workshop card tests

* workshop card tests for keypress

* max len lint fix

* requested changes

* tests fix

* added dialog with confirmation on editing workshop which draft already exists

* added translation to create workshop form and message bar on redirection

* req changes

* removed redundant click method

* removed code duplicate

* lint

* hotfix for details button

* requested changes

* workshop details test

+++ src/app/shared/components/workshop-card/workshop-card.component.ts
@@ -1,5 +1,7 @@
 import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
+import { Router } from '@angular/router';
+import { ENTER, SPACE } from '@angular/cdk/keycodes';
 import { Select, Store } from '@ngxs/store';
 import { Observable, Subject } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
@@ -21,7 +23,7 @@ import { WorkshopBaseCard, WorkshopDraft, WorkshopDraftCard, WorkshopProviderVie
 import { ImagesService } from 'shared/services/images/images.service';
 import { ShowMessageBar } from 'shared/store/app.actions';
 import { CreateFavoriteWorkshop, DeleteFavoriteWorkshop } from 'shared/store/parent.actions';
-import { DraftSendForModeration, UpdateWorkshopStatus } from 'shared/store/provider.actions';
+import { DraftSendForModeration, GetWorkshopDraftIdByWorkshopId, UpdateWorkshopStatus } from 'shared/store/provider.actions';
 import { RegistrationState } from 'shared/store/registration.state';
 import { MetaDataState } from 'shared/store/meta-data.state';
 import { FeaturesList } from 'shared/models/features-list.model';
@@ -76,9 +78,10 @@ export class WorkshopCardComponent implements OnInit, OnDestroy {
   private favoriteWorkshopId: string;
 
   constructor(
-    private store: Store,
-    private dialog: MatDialog,
-    private imagesService: ImagesService
+    private readonly store: Store,
+    private readonly dialog: MatDialog,
+    private readonly imagesService: ImagesService,
+    private readonly router: Router
   ) {}
 
   public get canOpenWorkshopRecruitment(): boolean {
@@ -108,31 +111,72 @@ export class WorkshopCardComponent implements OnInit, OnDestroy {
       });
   }
 
+  public onKeydown(event: KeyboardEvent, action: () => void): void {
+    if (event.keyCode === ENTER || event.keyCode === SPACE) {
+      action();
+      event.preventDefault();
+    }
+  }
+
+  public onEditKeydown(event: KeyboardEvent, id: string | undefined): void {
+    this.onKeydown(event, () => this.onEdit(id));
+  }
+
+  public onDeleteKeydown(event: KeyboardEvent): void {
+    this.onKeydown(event, () => this.onDelete());
+  }
+
+  public onLikeKeydown(event: KeyboardEvent): void {
+    if (this.role === Role.parent) {
+      this.onKeydown(event, () => this.onLike());
+    } else {
+      this.onKeydown(event, () => this.onOpenDialog());
+    }
+  }
+
+  public onDislikeKeydown(event: KeyboardEvent): void {
+    this.onKeydown(event, () => this.onDisLike());
+  }
+
+  public onEdit(workshopDraftId: string | undefined): void {
+    if (workshopDraftId) {
+      this.router.navigate(['create/draft', workshopDraftId]);
+    } else {
+      this.store.dispatch(new GetWorkshopDraftIdByWorkshopId(this.workshopData?.id));
+    }
+  }
+
   public onDelete(): void {
     this.deleteWorkshop.emit(this.workshopData);
   }
 
   public onLike(): void {
-    const param = new Favorite(this.workshopData.id, this.store.selectSnapshot(RegistrationState.parent).userId.toString());
-    this.store.dispatch([
-      new CreateFavoriteWorkshop(param),
-      new ShowMessageBar({
-        message: SnackbarText.addedWorkshopFavorite,
-        type: 'success'
-      })
-    ]);
-    this.isFavorite = !this.isFavorite;
+    if (this.role === Role.parent) {
+      const param = new Favorite(this.workshopData.id, this.store.selectSnapshot(RegistrationState.parent).userId.toString());
+      this.store.dispatch([
+        new CreateFavoriteWorkshop(param),
+        new ShowMessageBar({
+          message: SnackbarText.addedWorkshopFavorite,
+          type: 'success'
+        })
+      ]);
+      this.isFavorite = !this.isFavorite;
+    } else {
+      this.onOpenDialog();
+    }
   }
 
   public onDisLike(): void {
-    this.store.dispatch([
-      new DeleteFavoriteWorkshop(this.favoriteWorkshopId),
-      new ShowMessageBar({
-        message: SnackbarText.deleteWorkshopFavorite,
-        type: 'success'
-      })
-    ]);
-    this.isFavorite = !this.isFavorite;
+    if (this.role === Role.parent) {
+      this.store.dispatch([
+        new DeleteFavoriteWorkshop(this.favoriteWorkshopId),
+        new ShowMessageBar({
+          message: SnackbarText.deleteWorkshopFavorite,
+          type: 'success'
+        })
+      ]);
+      this.isFavorite = !this.isFavorite;
+    }
   }
 
   public onChangeWorkshopStatus(status: string, type: ModalConfirmationType): void {
